refactor: change structure code for AuthorList
