--- src/components/AuthorList/AuthorListCards/AuthorListCards.css
@@ -1,35 +1,0 @@
-/* исправить шрифт при resize */
-
-.img-fluid {
-    max-width: 100%;
-    max-height: 350px;
-}
-
-.photo {
-    width: 100%;
-    margin: 0;
-}
-
-.card {
-    border: none;
-    text-align: center;
-    display: flex;
-    align-items: center;
-}
-
-.card-body {
-    padding: 1rem;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-}
-
-*, ::after, ::before {
-    box-sizing: unset;
-}
-
-p {
-    width: 100%;
-}
