+++ src/components/AuthorList/index.tsx
@@ -1,1 +1,2 @@
 import './index.scss';
+export { default } from './AuthorListApp';
